fix(rmm-eac2): fix MISRA C:2012 Rule 10.7 in buffer.c and s2tt.c

+++ lib/realm/src/s2tt.c
@@ -1096,5 +1096,5 @@ unsigned long skip_non_live_entries(unsigned long addr,
 		}
 	}
 
-	return (addr + (i - index) * map_size);
+	return (addr + (unsigned long)(i - index) * map_size);
 }
